refactor to separate out bigger hooks
