+++ package.json
@@ -39,6 +39,8 @@
     "eslint-config-airbnb-typescript-prettier": "^3.1.0",
     "eslint-plugin-jest": "^24.0.2",
     "eslint-plugin-react-hooks": "^4.0.4",
+    "jest-websocket-mock": "^2.2.0",
+    "mock-socket": "^9.0.3",
     "msw": "^0.21.2",
     "prettier": "^2.1.2",
     "react-test-renderer": "^16.13.1",
